refactor: extract transcript field redaction helpers
